Added new attributes, data clean up and new methods - Added Get Cbft transactions method - Modified Mock data

+++ src/main/java/com/mvidyn/std/phub/ui/repository/ForeignPaymentRepository.java
@@ -3,5 +3,5 @@ package com.mvidyn.std.phub.ui.repository;
 import com.mvidyn.std.phub.ui.model.form.payment.ForeignPaymentForm;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ForeignPaymentRepository extends JpaRepository<ForeignPaymentForm, Integer> {
+public interface ForeignPaymentRepository extends JpaRepository<ForeignPaymentForm, Long> {
 }
